Removed spline interpolation
Spline interpolation was not working at all with MCMC. Instead
I am now modifying the phases and letting it properly integrate backwards.
The resultant waveform appears as if it were integrated forwards in time.

There is a small bug though, there is a slight phase shift between
the forwards integrated and backwards integrated waveform. Need
to fix this later. Just something to be aware of.